Tests/LibWeb/Text/expected/HTML/javascript-navigation-stays-in-same-process.txt

Hello world, this was printed by a JavaScript navigation!
